+++ crates/biome_service/src/file_handlers/vue.rs
@@ -11,12 +11,14 @@ use crate::WorkspaceError;
 use biome_formatter::Printed;
 use biome_fs::BiomePath;
 use biome_js_parser::{parse_js_with_cache, JsParserOptions};
-use biome_js_syntax::{EmbeddingKind, JsFileSource, TextRange, TextSize};
+use biome_js_syntax::{EmbeddingKind, JsFileSource, Language, TextRange, TextSize};
 use biome_parser::AnyParse;
 use biome_rowan::NodeCache;
 use lazy_static::lazy_static;
 use regex::{Match, Regex};
 use tracing::debug;
+
+use super::parse_lang_from_script_opening_tag;
 
 #[derive(Debug, Default, PartialEq, Eq)]
 pub struct VueFileHandler;
@@ -24,13 +26,7 @@ pub struct VueFileHandler;
 lazy_static! {
     // https://regex101.com/r/E4n4hh/3
     pub static ref VUE_FENCE: Regex = Regex::new(
-        r#"(?ixms)(?:<script[^>]?)
-            (?:
-            (?:(lang)\s*=\s*['"](?P<lang>[^'"]*)['"])
-            |
-            (?:(\w+)\s*(?:=\s*['"]([^'"]*)['"])?)
-            )*
-        [^>]*>\n(?P<script>(?U:.*))</script>"#
+        r#"(?ixms)(?<opening><script[^>]*>)\n(?P<script>(?U:.*))</script>"#
     )
     .unwrap();
 }
@@ -73,13 +69,17 @@ impl VueFileHandler {
     }
 
     pub fn file_source(text: &str) -> JsFileSource {
-        let matches = VUE_FENCE.captures(text);
-        matches
-            .and_then(|captures| captures.name("lang"))
-            .filter(|lang| lang.as_str() == "ts")
-            .map_or(JsFileSource::js_module(), |_| {
-                JsFileSource::ts().with_embedding_kind(EmbeddingKind::Vue)
+        VUE_FENCE
+            .captures(text)
+            .and_then(|captures| {
+                match parse_lang_from_script_opening_tag(captures.name("opening")?.as_str()) {
+                    Language::JavaScript => None,
+                    Language::TypeScript { .. } => {
+                        Some(JsFileSource::ts().with_embedding_kind(EmbeddingKind::Vue))
+                    }
+                }
             })
+            .map_or(JsFileSource::js_module(), |fs| fs)
     }
 }
 
@@ -120,11 +120,11 @@ fn parse(
     cache: &mut NodeCache,
 ) -> ParseResult {
     let script = VueFileHandler::input(text);
-    let language = VueFileHandler::file_source(text);
+    let file_source = VueFileHandler::file_source(text);
 
-    debug!("Parsing file with language {:?}", language);
+    debug!("Parsing file with language {:?}", file_source);
 
-    let parse = parse_js_with_cache(script, language, JsParserOptions::default(), cache);
+    let parse = parse_js_with_cache(script, file_source, JsParserOptions::default(), cache);
     let root = parse.syntax();
     let diagnostics = parse.into_diagnostics();
 
@@ -134,11 +134,7 @@ fn parse(
             root.as_send().unwrap(),
             diagnostics,
         ),
-        language: Some(if language.is_typescript() {
-            JsFileSource::ts().into()
-        } else {
-            JsFileSource::js_module().into()
-        }),
+        language: Some(file_source.into()),
     }
 }